Altered heroku population commands in sample_data.rake.

+++ lib/tasks/sample_data.rake
@@ -18,8 +18,14 @@ namespace :app do
   # Populates development data
   desc "Populate the database with development data."
   task :populate => :environment do
-    # Removes content before populating with data to avoid duplication
-    Rake::Task['db:reset'].invoke
+    if Rails.env.production?
+      # Heroku, reset migration
+      `heroku pg:reset DATABASE --confirm vox-vcp`
+      `heroku run rake db:migrate`
+    else
+      # Removes content before populating with data to avoid duplication
+      Rake::Task['db:reset'].invoke
+    end
 
     # INSERT BELOW
